+++ tests/Api/MergeRequestsTest.php
@@ -411,6 +411,68 @@ class MergeRequestsTest extends TestCase
     }
 
     #[Test]
+    public function shouldShowMergeRequestNoteAwardEmojis(): void
+    {
+        $expectedArray = [
+            ['id' => 1, 'name' => 'sparkles'],
+            ['id' => 2, 'name' => 'heart_eyes'],
+        ];
+
+        $api = $this->getApiMock();
+        $api->expects($this->once())
+            ->method('get')
+            ->with('projects/1/merge_requests/2/notes/3/award_emoji')
+            ->willReturn($expectedArray)
+        ;
+
+        $this->assertEquals($expectedArray, $api->showNoteAwardEmojis(1, 2, 3));
+    }
+
+    #[Test]
+    public function shouldShowMergeRequestNoteAwardEmoji(): void
+    {
+        $expectedArray = ['id' => 4, 'name' => 'sparkles'];
+
+        $api = $this->getApiMock();
+        $api->expects($this->once())
+            ->method('get')
+            ->with('projects/1/merge_requests/2/notes/3/award_emoji/4')
+            ->willReturn($expectedArray)
+        ;
+
+        $this->assertEquals($expectedArray, $api->showNoteAwardEmoji(1, 2, 3, 4));
+    }
+
+    #[Test]
+    public function shouldAddMergeRequestNoteAwardEmoji(): void
+    {
+        $expectedArray = ['id' => 4, 'name' => 'sparkles'];
+
+        $api = $this->getApiMock();
+        $api->expects($this->once())
+            ->method('post')
+            ->with('projects/1/merge_requests/2/notes/3/award_emoji', ['name' => 'sparkles'])
+            ->willReturn($expectedArray)
+        ;
+
+        $this->assertEquals($expectedArray, $api->addNoteAwardEmoji(1, 2, 3, 'sparkles'));
+    }
+
+    #[Test]
+    public function shouldRemoveMergeRequestNoteAwardEmoji(): void
+    {
+        $expectedBool = true;
+
+        $api = $this->getApiMock();
+        $api->expects($this->once())
+            ->method('delete')
+            ->with('projects/1/merge_requests/2/notes/3/award_emoji/4')
+            ->willReturn($expectedBool);
+
+        $this->assertEquals($expectedBool, $api->removeNoteAwardEmoji(1, 2, 3, 4));
+    }
+
+    #[Test]
     public function shouldGetMergeRequestParticipants(): void
     {
         $expectedArray = [
@@ -703,6 +765,36 @@ class MergeRequestsTest extends TestCase
         ;
 
         $this->assertEquals($expectedArray, $api->awardEmoji(1, 2));
+    }
+
+    #[Test]
+    public function shouldShowMergeRequestAwardEmoji(): void
+    {
+        $expectedArray = ['id' => 3, 'name' => 'sparkles'];
+
+        $api = $this->getApiMock();
+        $api->expects($this->once())
+            ->method('get')
+            ->with('projects/1/merge_requests/2/award_emoji/3')
+            ->willReturn($expectedArray)
+        ;
+
+        $this->assertEquals($expectedArray, $api->showAwardEmoji(1, 2, 3));
+    }
+
+    #[Test]
+    public function shouldAddMergeRequestAwardEmoji(): void
+    {
+        $expectedArray = ['id' => 3, 'name' => 'sparkles'];
+
+        $api = $this->getApiMock();
+        $api->expects($this->once())
+            ->method('post')
+            ->with('projects/1/merge_requests/2/award_emoji', ['name' => 'sparkles'])
+            ->willReturn($expectedArray)
+        ;
+
+        $this->assertEquals($expectedArray, $api->addAwardEmoji(1, 2, 'sparkles'));
     }
 
     #[Test]